workers/scan-worker/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
redis==5.0.1
nats-py==2.6.0
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.24.3
scipy==1.11.4
scikit-image==0.22.0
open3d==0.18.0
g2o-python==0.0.1
pillow==10.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
matplotlib==3.8.2
trimesh==4.0.5
